Stokhos:  Fix build errors in Stokhos due to recent changes in Kokkos
A recent Kokkos snapshot changed the assumed interface for Kokkos
execution policies.  Stokhos defines an execution policy for use with
Sacado::MP::Vector on Cuda to map Cuda threads to ensemble components
(and it much simpler to use than the normal team-based execution
policy), so it needed to be updated for these changes.