MueLu: changed requirements for Stratimikos adapter
We have two versions of a function to enable MueLu in Stratimikos
(Stratimikos::enableMueLu): non-templated and templated. The first one
is supposed to be used for Epetra, while the second one only for Tpetra.

The issue, though, is that there is nothing preventing running the
non-templated version for Tpetra, as that is a runtime decision.
However, the non-templated version assumes a specific set of parameters:
<double, int, int> with a serial node. This is exposed when trying to
compile MueLu with Tpetra without that instantiation, resulting in
undefined linkage with Tpetra's <double,int,int> objects.

The change made throws an exception if Tpetra is enabled, but without
that particular instantiation. That means that one is not able to run
both Epetra through non-templated enableMueLu and Tpetra through templated
enableMueLu at the same time.

I am not convinced that this is a right way to fix the issue.