Ifpack: Patched hypre interface
The hypre interface now rejects parallel distributions that will
make hypre unhappy.  (hypre wants A, b, and x to have the same row
map, and it also requires the maps to be linear and contiguous.)

We try to recover if the matrix has a bad distribution, but we
currently return an error code if the vectors have a bad distribution.