Tpetra::CrsMatrix::sumIntoLocalValues: No more dynamic alloc for GIDs
Users sometimes call sumIntoLocalValues when the CrsMatrix is globally
indexed.  This is a suboptimal decision, because it forces the matrix
to convert local indices into global indices.  However, this case does
happen, and has happened recently.

Before, in this rare case of calling sumIntoLocalValues with a globally
indexed matrix with a column Map, the matrix would allocate a temporary
array to convert the input local indices to global indices.  This
causes big slowdowns when calling this method in parallel on different
threads.  This commit removes the dynamic allocation by letting the
graph do the local-to-global conversion one index at a time.  It also
simplifies the method's implementation quite a bit and removes
redundant code.