Tpetra::CrsMatrix::replaceGlobalValues: Optimize and simplify.
The routine only does one global->local index lookup on the row
index now, instead of two (an error check was doing an extra
lookup).  It also no longer does an atomic update (the result of
replacing the same value concurrently by multiple threads is
undefined anyway, so there is no need for an atomic update).