Xpetra: Fix build of tests for when Tpetra Serial is disabled.
Xpetra recently changed in such a way that it would not build when
both Epetra and Tpetra were enabled, but OpenMP was turned on.  (This
turns off Tpetra's support for the Serial execution space by default.)
This commit fixes the build, though it does NOT fix Xpetra's failing
unit tests.