MueLu: remove damping factor from tests
This commits avoids an issue in Ifpack2, namely, the damping factor in
Relaxation (and probably other places) must be of type Scalar.

This means for a std::complex<double> matrix, one must set the damping
parameter type properly in an XML input deck.

In general, this is problematic.  The damping factor will probably be
double, but it doesn't have to be.

One possible solution is to use multiple catch/throws as is done in
RILUK and ILUT for level-of-fill.