Xpetra,MueLu: disabling all tests
Due to recent refactoring of ETI system, we have multiple failing tests
in these packages for different configurations. The team decided to
temporarily hide all tests behind a CMake configuration option, so that
people running checkin scripts do not trip on them. The tests will still
be available on the dashboard to keep track of the failures. The modus
operandi is to move the fixed tests out one by one.