Clean up some stk cmake files, move and remove a couple of dependencies.
Removed stk_search dependence on stk_unit_test_utils. (It was just a cmake file error,
not an actual dependency.)
Move stk_transfer/use_cases to stk_usecases/transfer so that the stk_transfer sub-package
doesn't need to depend on the things that the use-case programs depend on.

(From Sierra)