Xpetra: add full specialization of Xpetra::ThyraUtils
We need a full specialization of ThyraUtils on SC=double, LO=GO=int and NO=EpetraNode.
Thyra only supports Epetra but not Epetra64. Therefore, we do not add a specialization on
GO=long long.
The general implementation is supposed for Tpetra only. The Tpetra case in the specialization
for Epetra is guarded.