Tpetra: Check that index type in small dense LU kernels is signed
@trilinos/tpetra The new small dense LU factorization, solve, and
inverse kernels (for Tpetra::BlockCrsMatrix's blocks) implicitly
assume that the index type is signed.  This is because they have
loops that count down to zero.  Unsigned integers are always >= 0,
so a for loop like "for (IndexType i = N-1; i >= 0; --i) { ... }"
will never terminate if IndexType is unsigned.  Rather than fix
this, I decided just to make the new kernels do a static_assert to
check that the index type is a signed integer.