Sacado/Stokhos:  Add experimental view for ensemble scalar type.
This adds an experimental view specialization for the ensemble UQ scalar
type in stokhos.  It also fixes numerous issues with the Sacado FAD
specialization.  Everything appears to compile and run with the
experimental view for the ensemble scalar type all the way through the
Tpetra stack.  To facilitate this I added configure options for
selectively enabling/disabling the ensemble and PCE UQ scalar types.
Note the PCE scalar type has not even been started yet, so enabling it
and the experimental view will fail.  Also the changes in Trilinos
couplings haven't been tested yet for the experimental view because the
ensemble FENL requires both ensemble and PCE scalar types.  This code
also depends on some changes to the Kokkos experimental view that haven't
been pushed into Trilinos yet.