Piro: Generalize template signature for SolverFactory.
MueLu is used internally to the Piro_Rythmos solver. One needs
to specify the proper LocalOrdinal and GlobalOrdinal based on
the application (Albany) build that is desired. This commit generalizes
the template arguments in the Piro solver path to allow the correct LO and
GO based on the desired build.