Trilinos: Update muir testing
Fixed issue with boostlib not being found and disable STK on serial builds.

BoostLib just wasn't being set so the boost found was the system version
which is ancient. There was already a built boost on muir so pointing the
configure to that.

Also since STK hasn't built in serial for quite a while I've disabled it
on the serial builds on muir. The STK team has been made aware of the
issue and no fix seems forth coming so the red only makes it harded for
others to see new problems.