Stokhos:  Add support for LayoutStride for UQ scalar types.
Tpetra::BlockCrsMatrix appears to now use LayoutStride, which was not
supported by the view specializations for the UQ scalar types.  I added
some support for this (at least to get the Tpetra instantiations to
compile), but I haven't tested it functions correctly yet.  This should
fix the new nightly Cuda builds (hooray!)