Intrepid2: Fix a signed/unsigned conflict.
In
    intrepid2/src/Kokkos/Intrepid2_KokkosRank.hpp
there is the typedef
    typedef Kokkos::DefaultExecutionSpace::size_type index_type;
This resulted from the modification
    -typedef int index_type;
    +typedef Kokkos::DefaultExecutionSpace::size_type index_type;
made in commit e87d44d6.

size_type can be unsigned, which conflicts with the use of -1 in the pattern

    template<class DataT,index_type leftrank>
    struct RankSpec{};

    template<class DataT>
    struct RankSpec<DataT,-1>{
        ...
    };

Code does not necessarily actually break since the -1 becomes a
very large size_t, say (if size_type == size_t), which is fine if
it's just acting as a tag. However, compilers will warn about the
narrowing problem, -Wc++11-narrowing in clang.

An easy and correct solution is to modify
    template<class DataT,index_type leftrank>
    struct RankSpec{};
to be
    template<class DataT,int leftrank>
    struct RankSpec{};
and similarly for
    template<class Scalar,class ArrayType,index_type ArrayRank,bool isconstant>
    struct ArrayWrapper;
This commit does that.