Jamfile: Add rule to install Jamfile, Jamroot, boost-build.jam
The Jamroot, boost-build, and Jamfile files need to be installed in the
product so that when the tools are changed to use those files in their
new locations that user subroutines and plugins will work. (Currently
these files are still logically in the Nbtools directory via symlinks.)
This commit change the product Jamfiles to install themselves into the
end-user installation area in the product directory.

A rule to install the Jamroot and boost-build.jam files in the
installation root is added to Sierra/Jamfile by this commit.

This commit does nothing with the TPL Jamfiles. That will have to come
later.

(From Sierra)