SEACAS: Fix dependency issue with exodus in Trilinos build
Ioss was always stating a dependency for one of its libraries that is only
built when exodus is on. This was confusing cmake to put -l<lib name> on
the link line which of course was never resolvable. The fix was to change
the dependency to only add the library when exodus is enabled.

(From Sierra)