Tpetra: Remove redundant includes from Tpetra_ConfigDefs.hpp
@trilinos/tpetra This file includes way too many header files.
I would like to strip out most of them, but that may break
downstream code that expects those header files to be there.
However, what I _can_ safely do is remove header files that are
known to be redundant, since other header files include them.
This commit does that for a few header files that I know to be
redundant.