Fix a bug in mesh-mod related to ghost entity handling.
In this case a face and node were both shared and ghosted to the same
proc, then deleted from the owning proc. So the entities needed to be no
longer shared nor ghosted, and ownership moved to the new proc.

Fixing this allows the adagio remesh_donut_death test to run in
stk-parallel mode.

(From Sierra)