Phalanx: added hybrid assembly with AMT via Kokkos Task
Since AMT is experimental in Kokkos, it is always disabled with ifdefs
by default but can be enabled with:

-D Phalanx_ENABLE_KOKKOS_AMT:BOOL=ON

When enabled, AMT is implemented in a backwards compatible way so
codes will still compile. However, any evaluators that are used in a
Task parallel assembly call will fail at runtime unless they derive
from a new base class that supplies AMT required methods.