Tpetra::Map: Fix clone() for CUDA build
@trilinos/tpetra The Dashboard showed a build error for Map::clone
in CUDA builds.  It looks like Kokkos::deep_copy does not compile,
when the first (output) argument is a DualView on (Cuda,
CudaUVMSpace), and the second argument is a DualView on (Serial,
HostSpace).  The code in question affects lgMap_, a member of Map.
lgMap_ is a DualView that maps from local indices (implicitly
stored as all valid array indices of the Views) to global indices.

Tpetra assumes UVM anyway, so we don't need a DualView here.  I
changed the DualView to a View, and used create_mirror_view /
deep_copy instead of modify / sync for host code that fills it.

Build/Test Cases Summary
Enabled Packages: TpetraCore
Disabled Packages: FEI,STK,PyTrilinos,Intrepid2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_COMPLEX => passed: passed=83,notpassed=0 (0.49 min)
3) SERIAL_RELEASE => passed: passed=63,notpassed=0 (7.05 min)