MueLu: add a new routine ExtractCoordinatesFromParameterList to Utilities class
This routines checks the provided parameter list for an entry "Coordinates"
and returns it as an Xpetra::MultiVector. The scalar type of the coordinates
is expected to be either double or float. For Epetra the scalar type is expected
to be double, as Epetra has no support for other scalars.