Tpetra: Fix for CUDA run-time errors (host space not init'd) in test
@trilinos/tpetra Tpetra has a test for Kokkos::DualView; it exercises
subviews, in particular the ability to take degenerate (one or more
dimensions are zero) subviews.  The test has been failing in CUDA
builds, due to the host execution space not having been initialized.
This commit fixes that, by using the same strategy (initialize once
only, on demand, and register an atexit hook with each initialized
execution space to finalize it once) as the FixedHashTable test.