Tpetra: Remove dbg throw from generic LittleBlock/View fns (CUDA fix)
@trilinos/tpetra Tpetra::Experimental::AXPY implements matrix or vector
update ("a*x plus y," hence "AXPY") for LittleBlock, LittleVector, and
rank-2 or rank-1 Kokkos::View.  Tpetra::Experimental::COPY implements
matrix or vector deep copy for LittleBlock, LittleVector, and rank-2 or
rank-1 Kokkos::View.

In December, I specialized AXPY and COPY for LayoutRight rank-2 objects
(LittleBlock or Kokkos::View), as an optimization for non-GPU
architectures.  In order to tell whether my specialization or the
generic code was getting called, I inserted an exception throw in the
generic code as a temporary measure.  The generic code wasn't getting
called, so I continued with my business and forgot about the debug code.
Of course, with CUDA, LayoutLeft is the default Kokkos::View layout.
(That's a big point of Kokkos: it chooses the best data layout for the
architecture.)  Since I only had a specialization for LayoutRight, the
CUDA build invoked the generic code, which resulted in the exception
being thrown (with cute debugging message "OH HAI") and the test
failing.

This commit fixes the CUDA build both for correctness and performance.
First, I took the debug code (that throws the exception) out of the
generic AXPY and COPY implementations.  Second, I added specializations
of AXPY and COPY for LayoutLeft rank-2 objects, that optimize for
performance (by flattening the nested loop into a single loop).