MueLu: added smoothers and coarse solver reuse ("reuse: type" = "S")
For any reuse type (tP, RP, RAP, emin), the finest matrix is assumed to
be structurally unchanged. Thus, it makes sense to reuse some smoother
data. The way to approach that is to keep the constructed smoothers on
the level, and call Setup() for consecutive setups. This way, as long as
the corresponding SmootherPrototype handles consecutive Setup() in a
proper reuse manner, we should be good.

If reuse is wanted only for the finest level smoothers, put the option
into an appropriate level sublist "level X".

For now, we only reuse coarse solver if it is not a direct
solver, so that it still uses TrilinosSmoother. I'll code the
DirectSolver later.