Tpetra::Map::clone: Fix for CUDA build
@trilinos/tpetra Tpetra::Map::clone was invoking deep_copy from a
Kokkos::View<GO*, D_in> to a Kokkos::View<GO*, D_out>, where D_in is the
input Device type and D_out the output Device type.  This is incorrect
if the two Devices' default View layouts differ, which is indeed the
case if one of the Devices is a host Device (OpenMP, Serial, or Threads)
and the other is a CUDA Device.

This commit fixes the layout of the View in question to be LayoutLeft.
Either LayoutLeft or LayoutRight would be correct.  I picked LayoutLeft
for debugging, because it is _not_ the default layout on non-CUDA
Devices.  This should help catch errors involving incorrect assignment
or copying from a View with default layout to this View.