Anasazi: Fix (release only) build warnings in Epetra LOBPCG example
@trilinos/anasazi The 'assert' macro is only defined in a debug build.
In a release build, it disappears.  The Epetra LOBPCG example in
question uses the 'info' int return value of various Epetra methods only
for 'assert(info != 0)'.  This means that in a release build, the 'info'
return values are never used.

This commit replaces 'assert(info != 0)' with a
TEUCHOS_TEST_FOR_EXCEPTION test.  That test is always defined, and it
achieves the same goal of a run-time test.