Anasazi: Disable some TraceMin* tests/examples if TSQR enabled
@trilinos/anasazi @trilinos/tpetra

Some of Anasazi's TraceMin* tests and examples use a custom
specialization of MultiVecTraits for their own multivector type.  The
specialization omits the TSQR adapter typedef.  This causes build errors
when TSQR is enabled in Anasazi.  TSQR is disabled by default, so
implementers tend not to see that they need the TSQR adapter in their
MultiVecTraits specialization.

The complete fix would be to implement a TSQR adapter for the custom
multivector type, and then to make this the 'tsqr_adaptor_type' typedef
in the MultiVecTraits specialization.  (This could even just be a stub
that compiles, but whose methods all throw std::logic_error; Thyra's
TSQR adapter does this currently.)  For now, I'm just disabling the
tests and examples in question when TSQR is enabled.  Since TSQR is
disabled by default, the tests and examples will still normally get
exercised.