Fixed gcc debug build error
The updated gtest macro definition of EXPECT_TRUE() doesn't play nice
with this one usage of accessing a bool inside a template instantiation
when building with (at least) gcc-4.7.2 debug.  Not sure why.
Comparing the bool to true makes it so that the macro doesn't have to
resolve the template data member directly.

(From Sierra)