Removed BOOST_GPU_ENABLED usage from stk
This is a boost-specific macro definition that defines a function on the
host or device for a GPU.  If a function is declared with this modifier, then
anything it calls must also have this modifier.  This is leading to
build problems on some of the new NGP platforms due to inconsistent
usage in both STK and Boost.  Since this isn't the official path forward
for NGP support anyway, just get rid of it.  There is an equivalent
KOKKOS_INLINE_FUNCTION macro that we may end up using, but we will cross
that bridge when we get to it.

(From Sierra)