MueLu: adapt navier-stokes tests in MueLu test folder
These tests run only with Epetra as they read in a matrix in matrix-market format
and split it into a 2x2 block matrix. The tests will be ported as soon as the splitting
routine is Xpetra-fied in context of the Drekar project. This commit just should fix
the build errors in the dashboard.