IOSS: Fix issue with entity_processor on side commset overflowing
The previous behavior was that the entity_processor map was a pair consisting of entity id and proc
For a "side" commset, the entity id was 10*element_id+local_side_id.  If the model had more than
214 million elements, or if the global element id max was larger than this value, then
multiplying by 10 would overflow an integer.  This was trapped, but didn't do much good
for the analyst trying to run an analysis.

Changed the entity_processor field for a "side" commset to a triplet consisting of
element_id, local_side, processor and now there is no overflow.

However, the issue is that the field size has changed, so any applications using
this field will no longer work.  This shouldn't be an issue since there was also
a bug discovered in which the output of the side commset entity processor information
was not correctly mapping from global id to local id, so any code using this should
have already been crashing...

checked all sierra-based codes to make sure they work correctly; aleph works;
Salinas does not use this; need to check others also...

(From Sierra)