Intrepid MiniTensor: Changed implementation of subtensor.
The recursive implementation of the determinant function for N > 3 was
throwing a warning in Sierra, and in Sierra warnings are treated as errors.
This is because of some template metaprogramming used to figure out
the correct dimensions of the subtensors used in the recursion.

The recursion together with the compile-time dimension calculation are no
longer needed as statically allocated tensors can now change
dimension as long as it does not exceed the capacity specified at compile
time.

We keep the recursive algorithm but eliminate the template used to
calculate the dimension. This avoids the warning and also simplifies
the implementation a bit.