MueLu: use GetOstream instead of cout in MLParameterListInterpreter
In addition, it fixes ML interface test that fail due to the presense of
extra output when ML is disabled.

The work is not complete at this point, as one needs to deal with
ML2MueLuTranslator class to use the GetOStream mechanism. It is not
possible at this point as most of its functions are static, and thus
even if we inherited it from the BaseClass we still would not get the
access to GetOStream.

Build/Test Cases Summary
Enabled Packages: MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=21,notpassed=0 (15.88 min)