More fixes for failing stk unit tests and aria regressions tests when check_exposed_boundary_sides() is enabled with partial graph edge coincidence detection.
The last two cases had to do with:

1) Usage of get_ordinal_and_permutation() was not clearly defined
because some places cared about filtering negative permutations and
some did not.

2) Creation of exposed sides in the case of a refined element where
the parent element is designated as 'air' and the children are skinned
on a parallel boundary that is connected to elements to be
skinned. This needed to be augmented with checks on the coincident
graph because the old fully coincident code did not detect the partial
coincidence of the children with the parent.

(From Sierra)