Panzer:  Rework how tangent parameter names are passed.
Previously I was relying on the list of tangent parameter names to be
passed through the parameter list for each equation set.  The problem
with that is it requires a few small changes to each equation set to
enable tangent sensitivities.  Now one can set the tangent parameter
names through a virtual function on the equation set, and the physics
blocks can take this list as an argument to be passed to the equation
sets.  This allows the parameter names to be initialized in a simpler
manner with no changes to equation sets.