Added experimental replacement for create_faces() based on ElemElemGraph
This has the benefit of simplicity and significant code re-use with all
of the other skinning capabilities.  There is the potential for a
performance win here as well, after some future tuning effort.

This is protected with the stk::mesh::experimental namespace, so is not
yet deployed into production.  Some unit testing does exist, however.

Collaborators: Alan Williams, Tolu Okusanya

(From Sierra)