LOCA:  Attempting to fix failing tests on 12.6 release branch.
The dashboard is showing repeated failing tests for the householder
bordered solver unit tests.  I can't reproduce these, but I am
suspicious that it is the use of random bordering matrices.  I think it
is possible different matrices are being generated on different
processors.  This commit attempts to fix this by setting the seed
explicitly.