Added the BelosUtils file that contains
std::pair< std::string, int >
Belos_Translate_from_Aztec_Params(const int * aztec_options,
				  const double * aztec_params,
				  Teuchos::ParameterList &tpl )

and a test packages/belos/test/Util/Aztec_to_Belos_param.cpp that exercises it.
This test is _not_ an exaustive test, it only touches a portion of the aztec options
and one parameter.

The returned pair contains a std::string with a error or warning text, and an int for the error bit mask, so the return can have both a WARN and an ERROR on the same call.
 enum XLateStatus {
    OK    =0x00,
    WARN  =0x01,
    ERROR =0x02};