MueLu: fix failing test emin1.xml
In MueLu_CGSolver_def.hpp around line 127, there was this code:

    R->scale(-one);
    if (useTpetra)
      R->fillComplete(R->getDomainMap(), R->getRangeMap());

yet prior to this, R->resumeFill() is always being called.  For the test
case failing, linAlgebra=Epetra.  Epetra itself doesn't have a
resumeFill() method, but the Xpetra_EpetraCrsMatrix does.  So I believe
it is always safe to call resumeFill(), since the cost is just
checking or toggling the value of a boolean flag.