MueLu: attempt to fix Hierarchy::Write test
Add optional 3rd argument to Hierarchy::Write method for file suffix.
In the unit test, this is set to be the scalar type. The theory is that
for dashboard tests with multiple Scalar instantiations of this test, a
test with Scalar type A could try to read in the results of the test
with Scalar type B, simply because the test with type B overwrote A's
output matrix file before the A file could be read back in.  A better
solution would be to write to a file stream, but this would involve
writing new interfaces to Epetra's file I/O capabilities.

This commit also adds code to the Write test to remove the temporary
files created by the unit test.

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=86,notpassed=0 (13.11 min)