Belos: Implementation of a status test which allows to check the norm of a subpart of the residual vector
This is useful, e.g., for multiphysics probelms. For example, in a Navier-Stokes
example it is possible to define separate checks for the pressure part only or for
the velocity part. This gives a better control over the quality of the solution.

Note: this is only working for Thyra so far. The file contains a specialization for Thyra operators/multivectors