STK: Fix integration issue
Due to Sierra removing history a small anomaly was introduced during
the integration. Trying to pull in the proper commit where the missing
change was introduced causes significantly more problems since the
"top skim" commit that introduced the change is before the last integration
which results in many redundant changes being pulled in. Since the missing
changes are very small it is easier to just manually update the history.