Fix ArrayView::getConst() for raw ptrs (Trilinos #184)
This fixes a defect that resulted in an invalid "dangling reference" error.
Fixing this was a oneline change.  See the new unit test for details.

I did not add a unit test for Tpetra itself becuase it is redundant but I
verified manually that it fixes that use case (see Trilinos #184).

Fixes #184.