Update to new modules
The new modules define GCC_PATH but *not* GCCROOT anymore.  The old modules
did not define GCC_PATH.

For some reason, I also had to add the ldl library as well for shared builds.

I can't get anything to link with the new libraries on Hansen but at least we
make it through configure okay.  The current link error that I am seeing is:

/home/projects/x86-64-haswell/bintuils/2.25.0/lib/../x86_64-unknown-linux-gnu/bin/ld: \
/home/projects/x86-64-haswell-nvidia/hdf5/1.8.15/openmpi/1.10.0/gcc/4.8.4/cuda/7.5.18/lib/libhdf5.a(H5PL.o): undefined reference to symbol 'dlclose@@GLIBC_2.2.5'