Add TPL enables/disables, default search for static (Trilinos #172)
After I was given Matt's configure script, I could see all of the TPL enables
and disables.  I added these same enables/disables to the module (set as
defaults, not hard overrides).  I saw that the TPL HDF5 was explicitly
enabled.

I also made the search for TPL libs look for and only use static libs since
that is all there is on this machine.

With this, I was able to enable all of SEACAS and STK and pass all the tests
(except for the perpetual failing test
STKUnit_tests_util_parallel_UnitTest_MPI_4).