Tpetra: Add findRelOffset to fix #205
Tpetra::Details::findRelOffset is a new function (free function, not
method) that finds the relative offset of an index in a row of a sparse
graph or matrix, given an array of column indices in that row.  The
array may be either a rank-1 Kokkos::View or a raw 1-D array.  The
function lives in tpetra/core/src/Tpetra_Util.hpp.

This fixes #205.  I mean for the new function to replace or implement
Tpetra::CrsGraph::find{Local,Global}Index.

The function optimizes for the case where the input array to search is
sorted, by using binary search in that case.  This should make its
performance comparable to that of Tpetra::CrsGraph::findLocalIndex or
Epetra_CrsGraph::FindMyIndexLoc.

I also added a test for this function.  The test builds and passes.
Note that it carefully exercises both the sorted and unsorted cases.

Next step, not yet implemented is to optimize for the case of short
rows, by using linear search in that case.  This should actually improve
on both Epetra and Tpetra, and possibly address #118.