ML: add check for failed Ifpack smoother ctor
This commits checks whether the Ifpack_Preconditioner returned from
the Ifpack::Factory is null, in the case that ML is using Ifpack as
a smoother.  If the Ifpack pointer is null, ML now throws an exception.
(Ifpack's parameter validation is weak, so it's possible to get a null
Ifpack object back.)

Build/Test Cases Summary
Enabled Packages: ML
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=173,notpassed=0 (17.83 min)
1) SERIAL_RELEASE => passed: passed=171,notpassed=0 (11.36 min)