Trilinos: Make Trilinos-test use system cmake
The download script is failing to download a new cmake and this is
actually preventing the jobs from running. Other machines that use
this same script are not showing this issue. trilinos-test is very
old (RHEL5) and is going to be retired soon. There happens to be a
2.8.12 installation of cmake in the default path for hudson so it
makes sense to just switch trilinos test to use that for now to get
some testing from it instead of debugging an issue for a machine that
will soon be retired.