Fix to get ghosted entities promoted to shared
This promotes ghosts to shared under the StkTransitionBulkData custom
modification_end for entity creation (which is used inside the
Fmwk_Entity_Creation_Helper).  Calling the mod_end for just the rank
being created does not give lower-ranked entities a chance to be
promoted to shared, so this adds all lower-ranked entities to the list
of those to be processed.

Note that this is still an issue in pure STK apps, which will require
more effort to pull off without a notable performance hit (because STK
doesn't use DistributedIndex anymore and we would have to set up
additional rounds of communication to synchronize this information).

A disabled unit test has been added as a reminder that this is not a
working capability in STK.

(From STK/SEACAS)