IOSS: Handle api byte size settings correctly
If api byte size set to 64 bits and then to 32 bits, Ioss was then
using 32 bits, but the message didn't make it down to exodus library,
so it was still sending 64 bits.

Modified so that a call to set the api byte size propagates down
to the exodus library so Ioss and exodus are in sync.

Fixes #14718

(From STK/SEACAS)