MueLu: added no reuse to reuse loop in StandardReuse
This is mostly for extra information. For instance, running 2D laplacian
produces

Setup #1: no reuse                  0.02187 (1)
Setup #2: reuse none                0.01018 (1)
Setup #3: reuse smoothers           0.01025 (1)
Setup #4: reuse tentative P         0.006917 (1)
Setup #5: reuse smoothed P and R    0.002708 (1)

This means that about 50% of no reuse time is actually spent in the
interpreter, and the real construction time take 50%. This also shows
that there is no reuse happening in smoothers which could not be seen if
one only compare #1 and #3.