MueLu: add MueLu/Xpetra version of StatusTestGenResSubNorm.
This status test class allows to check norms of residual sub-vectors. It
is a specialization on the Xpetra operators and only supposed to work with
an underlying Xpetra::BlockedCrsMatrix object.
Please note, that the Belos interface in MueLu uses Belos::OperatorT operator
instances which internally store the corresponding Xpetra::Matrix object. Therefore,
the specialization is on Belos::OperatorT rather than Xpetra::BlockedCrsMatrix.
We throw, if Belos::OperatorT only encapsulates an Xpetra::Matrix object, but no
BlockedCrsMatrix object.