Scripts: add export for environment variables to SEMS based script
This now sets CUDA_LAUNCH_BLOCKING and CUDA_MANAGED_FORCE_DEVICE_ALLOC
two commonly necessary settings. It also unsets OMPI_CXX when not building
for Cuda. This is a common error, which lets one compile with NVCC if one
forgets to unset the variable after previously building in the same shell
for Cuda.