MueLu: fix Amesos2 solver for maps with gaps
If our maps contain a gap in them, for instance GIDs are [0,..., 100,
10000, ..., 10010], then Amesos2 breaks down.

This patch partially addresses the problem.  The approach we take is to
construct a second system with maps replaced by their continuous
versions.

NOTE: for the moment, this functionality works for selected limited scenarios.

Build/Test Cases Summary
Enabled Packages: MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=22,notpassed=0 (6.00 min)
Other local commits for this build/test group: 6cb0f06