Reduce stk::mesh::Entity size from 64bit to 32bit and ConnectivityOrdinal from 32bit to 16bit.
This does not limit the number of global entities that can exist in
stk-mesh, but it limits the number of entities per MPI process to 4
Billion or less, and it limits the number of connectivities per entity
to 65535 or less.
These are compile-time options that are turned on by the macros
STK_32BIT_ENTITY and STK_16BIT_CONNECTIVITY_ORDINAL

To revert to 64bit Entity and 32bit ConnectivityOrdinal, make sure
these macros are not defined.

(From STK/SEACAS)