PARSER: detect aprepro parsing errors
The infrastructure to detect aprepro parsing errors
is in place, but the final RuntimeDoomed call is commented
out currently.  Will either enable for all, or will add
a command-line switch which will either disable or enable.

This detects all aprepro parsing errors and if there are
any stops execution.  PRoblem is with false-positives such as
a line containing "# rhoc_{v}" as a comment triggers an
undefined variable errror for "v" even though this is just some
leftover LaTeX formatting.

On the current nightly regression test suite, enabling this option
detected three false positives and one true parse error...

Leaving disabled for now while investigating options and discussing
with relevant parties.

This is for ticket 13589

(From STK/SEACAS)