Panzer:  Added Multi-variate Parameter evaluator.
This is just like the panzer::Parameter evaluator, except it computes
the parameter value as a sum of n independent parameters.  This is
useful for scaling studies where you need to easily vary the number of
parameters.

I also touched CMakeLists.txt to force a reconfigure to pick up the new
files.