Tpetra::DistObject: Get rid of exports_old_ (Teuchos::Array)
@trilinos/tpetra DistObject no longer has exports_old_, a
Teuchos::Array<packet_type>.  Now it uses exports_ (a Kokkos::DualView)
for all communication operations.

In order for this to work, I had to make a backwards-incompatible
interface change to an internal function, packAndPrepareWithOwningPIDs.
That function is used mainly in sparse matrix-matrix multiply.  As a
result of the change, I had to fix the test (ImportExport2).

This relates to debugging #227.