PyTrilinos: Changes to get LOCA to import into Python 3
The main problem is that general 'import' statements do not do local
imports within a package in Python 3. To get local imports, you need
to do a 'from . import ...', but swig does not generate these
statements properly, even with the -relativeimport command line
option.

The fix is to find the current directory from the module's __file__
attribute, and add it to sys.path within the Python code of the
module. This commit represents an incomplete reworking of the LOCA
SWIG interface files to suport this.