Panzer:  Remove reverse communication in DOFManager::buildGlobalUnknowns.
This helps with issue #227.  Tpetra's import/export operations don't
appear to function correctly when using reverse communication (import
with exporter or export with importer) on Cuda.  As a workaround, I
removed the two instances of this communication pattern in
DOFManger::buildGlobalUnknowns.  This gets all of the panzer tests that
I can enable without STK to pass for me on Cuda, and so lets Panzer work
past this issue until it can be resolved in Tpetra.  Of course this
won't fix any other reverse communication calls elsewhere.