Tpetra::{MultiVector,Vector}: Implement describe() correctly
@trilinos/tpetra MultiVector and Vector now implement describe()
correctly, in that now only Process 0 of the object's communicator
prints to the given output stream.  The changes have the side benefit of
making output deterministic, and reducing code duplication between
MultiVector and Vector (they both defer to a common output method now).

I implemented this change on a different computer to help debug #227,
but then fixed #227 before I had a chance to merge it in.  I didn't want
it to be lost, so I'm adding it now.