Tpetra: Always assume that MPI can access CUDA device memory
@trilinos/tpetra See #227 discussion.  MPI implementations have been
able to read from and write to CUDA device memory for many years now.
This has nothing to do with GPU Direct or UVM.  It only requires that
the MPI implementation be able to tell whether a given pointer points
to host memory or CUDA device memory.  The MPI implementation may
choose to use any available mechanism to access that memory.  That
means Tpetra does not have to worry about it.

This commit includes the following changes:

  - I removed the Tpetra_ENABLE_MPI_CUDA_RDMA CMake option, and its
    associated C preprocessor macro.  (The name is not accurate anyway;
    this has nothing to do with RDMA.)

  - Tpetra::Distributor still accepts the "Enable MPI CUDA RDMA
    support" ParameterList option, but it must be true, else
    setParameterList throws an exception.

  - I removed the alternate code path from Distributor's
    doPostsAndWaits and doReversePostsAndWaits (both the 3-argument
    and 4-argument versions) that copies message buffers to/from host.
    That code path wasn't working anyway.  See #227 discussion.