Tpetra: Fix #331 (MatrixMatrix unit tests for CUDA)
@trilinos/tpetra Fix sparse matrix-matrix multiply unit tests for CUDA.
With Teuchos_ENABLE_DEBUG=ON, the tests were failing with a dangling
Teuchos::ArrayRCP reference.  The issue was that the original author of
the unit tests wrote code like this:

for (Teuchos::ArrayView<const GO>::iterator it = rowMap->getNodeElementList().begin ();
     it != rowMap->getNodeElementList().end(); ++it) {
 // ...
}

Tpetra::Map::getNodeElementList() returns Teuchos::ArrayView<const GO>.
The above code may be ugly, but it's technically correct; the first call
of getNodeElementList() creates the list of indices if necessary, then
caches it for later calls.

For CUDA, this method returns a nonowning view of the HostMirror of the
Map's data.  Map keeps a persistent HostMirror around, so this is
semantically correct.  However, this code appears to break
Teuchos::ArrayView debug tracking of dangling references.  I rewrote the
code to look like this:

Teuchos::ArrayView<const GO> gblRows = rowMap->getNodeElementList ();
for (auto it = gblRows.begin (); it != gblRows.end (); ++it) {
  // ...
}

and the tests passed.  This is better code anyway, because it keeps side
effects out of the loop bounds.

Build/Test Cases Summary
Enabled Packages: TpetraCore
Disabled Packages: FEI,PyTrilinos,Moertel,STK,SEACAS,ThreadPool,OptiPack,Rythmos
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG => passed: passed=90,notpassed=0 (3.50 min)
3) SERIAL_RELEASE => passed: passed=68,notpassed=0 (2.53 min)
Other local commits for this build/test group: 7feeeeb