Sacado:  Reduce build time for Sacado Kokkos tests.
It isn't really necessary to test all four Fad families.  It also isn't
necessary to test the default layout versus explicitly specifying a
layout.  This reduces the build time for these tests substantially.  I
also removed explicitly computing the number of threads in the
Threas/OpenMP tests, as it is unnecessary.