PyTrilinos: Update NOX.Abstract typemaps
I was getting a memory leak warning when I ran NOX example
tests. These were related to the fact that abstract classes (anything
in the NOX.Abstract namespace) are typically not useful to a Python
user, and so when these are returned from a method, say, they need to
be downcast to a concrete class. In the case of a NOX.Epetra.Vector,
it can be further converted to an Epetra.Vector.

I had made an incorrect assumption about NOX.Abstract.Vectors (namely,
that they are stored within PyTrilinos with Teuchos::RCPs, and this
led to a return type that SWIG did not fully understand, and so did
not have a destructor for, and so warned about a memory leak.

These changes fix the appropriate typemaps and eliminate the memory
leak warning.