Fixed OctTree search key indexing error
The OctTree implementation includes an OctTreeKey that holds up to 16
4-bit integers packed into two 32-bit unsigned ints, representing which
octant of the branch points at each depth of the tree we should follow
to get to the desired value.  Storing and retrieving values for each
of the 16 depths involves doing some calculation to see which of the two
unsigned ints holds the desired value, and then bit-shifting that unsigned
int to line up with a 4-bit mask to extract the desired octant value.

For some reason, the convention for octant numbering is to use the
integer range 1:8, rather than 0:7.  This lead to an error with the
modulo operator when retrieving the value for depth 8, which caused the
32-bit unsigned int to be shifted by 32 bits (rather than the desired 0
bits).  According to the C++ spec, this is undefined behavior.  We got
lucky on many platforms/compilers/optimization levels that the end
result is usually a shift of 0 bits, which is what we wanted.  On (at least)
the intel-14 compiler at -O3 optimization, the net result was also a shift
of 0 bits except the mask fell off and let a huge number through.

This commit fixes these errors.  The OctTreeKey::set_value() function that
is intended to paste in raw data to the key was also rewritten to do
precisely what it was intended to do, rather than doing all kinds of bit
shifting and masking to copy the two unsigned ints into place, 4 bits
at a time.

The unit testing that was in place only tested up to a tree depth of 4.
This was extended up to the maximum 10 supported by the tree data
structure (although the key implementation can technically go up
to 16).

(From STK/SEACAS)