Tpetra: Don't rely on UVM for row offsets in BlockCrsMatrix
@trilinos/tpetra Tpetra::Experimental::BlockCrsMatrix currently makes
two assumptions that we need to remove for full native CUDA support:

  1. It assumes that it is possible to assign the local graph's
     row_map to ptr_, which is a HostMirror.  This only works on CUDA
     because Tpetra currently uses Kokkos::CudaUVMSpace as its memory
     space, and the HostMirror of a CudaUVMSpace View is also in
     CudaUVMSpace.

  2. It assumes that ptr_ is accessible on both host and device.
     This works for the same reason that (1) above works.

This commit removes these assumptions.  ptr_ becomes ptrHost_, which
is a HostMirror.  However, BlockCrsMatrix now only accesses the
entries of ptrHost_ in host code.  BlockCrsMatrix does not need to
store the device version of ptrHost_, because that already exists in
the graph (as graph.getLocalGraph().row_map).

These changes will make the full fix for #178 easier.