Tpetra: Don't rely on UVM for local column indices in BlockCrsMatrix
@trilinos/tpetra Tpetra::Experimental::BlockCrsMatrix currently
assumes that the entries of the array of column indices ind_ are
accessible on both host and device.  This implicitly assumes UVM.

This commit removes that assumption.  ind_ stops being a raw pointer,
and becomes a HostMirror (Kokkos::View) of the local graph's (device)
View of column indices.  BlockCrsMatrix now only accesses the entries
of indHost_ in host code.  BlockCrsMatrix does not need to store the
device version of indHost_, because that already exists in the graph
(as graph.getLocalGraph().entries).

These changes will make the full fix for #178 easier.